WebresourceServer is back in business
