+++ osgi.enroute.web.simple.provider/src/osgi/enroute/web/server/provider/BundleMixinServer.java
@@ -27,7 +27,7 @@ import osgi.enroute.webserver.capabilities.*;
 		service = { ConditionalServlet.class }, 
 		immediate = true, 
 		property = {
-				"service.ranking:Integer=1001", 
+				"service.ranking:Integer=1004", 
 				"name=" + BundleMixinServer.NAME}, 
 		name = BundleMixinServer.NAME, 
 		configurationPid = BundleMixinServer.NAME,
