+++ osgi.enroute.web.simple.test/bnd.bnd
@@ -18,7 +18,8 @@ Conditional-Package: \
 	osgi.enroute.base.api;version=1.0,\
 	osgi.enroute.web.simple.provider;version=latest,\
 	biz.aQute.bndlib;packages=*, \
-	osgi.enroute.junit.wrapper;version=4.12
+	osgi.enroute.junit.wrapper;version=4.12,\
+	net.sourceforge.htmlunit;version=2.15
 
 -runrequires: \
 	osgi.identity;filter:='(osgi.identity=osgi.enroute.web.simple.test)',\
@@ -27,6 +28,7 @@ Conditional-Package: \
 -runbundles: \
 	biz.aQute.bndlib;version='[3.0.0,3.0.1)',\
 	com.springsource.org.json;version='[1.0.0,1.0.1)',\
+	net.sourceforge.htmlunit;version='[2.15.0,2.15.1)',\
 	org.apache.commons.fileupload;version='[1.3.1,1.3.2)',\
 	org.apache.commons.io;version='[2.4.0,2.4.1)',\
 	org.apache.felix.configadmin;version='[1.8.6,1.8.7)',\
@@ -45,10 +47,29 @@ Conditional-Package: \
 	osgi.enroute.configurer.simple.provider;version=snapshot,\
 	osgi.enroute.dtos.bndlib.provider;version=snapshot,\
 	osgi.enroute.executor.simple.provider;version=snapshot,\
+	osgi.enroute.github.angular-ui.webresource;version=snapshot,\
+	osgi.enroute.google.angular.webresource;version=snapshot,\
 	osgi.enroute.hamcrest.wrapper;version=snapshot,\
 	osgi.enroute.junit.wrapper;version=snapshot,\
 	osgi.enroute.logger.simple.provider;version=snapshot,\
+	osgi.enroute.stackexchange.pagedown.webresource;version=snapshot,\
+	osgi.enroute.twitter.bootstrap.webresource;version=snapshot,\
 	osgi.enroute.web.simple.provider;version='[1.3.3,1.3.4)',\
+	osgi.enroute.web.simple.provider;version='[1.3.4,1.3.5)',\
 	osgi.enroute.web.simple.provider;version=snapshot,\
 	osgi.enroute.web.simple.test;version=snapshot,\
 	osgi.enroute.webconsole.xray.provider;version=snapshot
+
+Require-Capability:		\
+	osgi.enroute.webresource; \
+		filter:='(osgi.enroute.webresource=/github/angular-ui)'; \
+		resource:List<String>='ui-bootstrap-tpls.js', \
+	osgi.enroute.webresource; \
+		filter:='(osgi.enroute.webresource=/google/angular)'; \
+		resource:List<String>='angular.js,angular-resource.js', \
+	osgi.enroute.webresource; \
+		filter:='(osgi.enroute.webresource=/stackexchange/pagedown)'; \
+		resource:List<String>='enmarkdown.js', \
+	osgi.enroute.webresource; \
+		filter:='(osgi.enroute.webresource=/twitter/bootstrap)'; \
+		resource:List<String>='css/bootstrap.css'
