+++ osgi.enroute.web.simple.test/src/osgi/enroute/web/simple/test/BundleConfig.java
@@ -1,7 +1,9 @@
 package osgi.enroute.web.simple.test;
 
 import osgi.enroute.configurer.api.RequireConfigurerExtender;
+import osgi.enroute.webserver.capabilities.RequireWebServerExtender;
 
+@RequireWebServerExtender
 @RequireConfigurerExtender
 public interface BundleConfig
 {
