+++ osgi.enroute.web.simple.test/src/osgi/enroute/web/simple/test/BundleMixinServerTest.java
@@ -32,14 +32,16 @@ public class BundleMixinServerTest extends TestCase {
 		super.tearDown();
 	}
 
+	/**
+	 * This will serve the "debug" version of the WebServer INDEX page.
+	 */
     public void testBundleMixinServer() throws Exception {
         assertValid("enRoute INDEX", 200, "http://localhost:8080/index.html");
     }
 
     /**
      * If only the hostname is provided (no slash), then the default action by the RedirectServlet
-     * will append "/index.html". In debug mode, WebServer will call upon the cache to serve up the 
-     * contents of the static/debug/index.html file.
+     * will append "/index.html". This will serve the "debug" version of the WebServer INDEX page.
      */
     public void testHostNameOnly() throws Exception {
         assertValid("enRoute INDEX", 200, "http://localhost:8080");
@@ -47,8 +49,7 @@ public class BundleMixinServerTest extends TestCase {
 
     /**
      * If the request is for an empty path ("/" only), then the default action by the RedirectServlet
-     * will append "index.html". In debug mode, WebServer will call upon the cache to serve up the 
-     * contents of the static/debug/index.html file.
+     * will append "index.html". This will serve the "debug" version of the WebServer INDEX page.
      */
     public void testEmptyRootPartition() throws Exception {
         assertValid("enRoute INDEX", 200, "http://localhost:8080/");
